fixed stupid return 0 bug :)

+++ BasicOpenGLEngine/BasicOpenGLEngine/PointLight.cpp
@@ -8,7 +8,7 @@ PointLight::PointLight()
     exponent = 0.0f;
 }
 
-PointLight::PointLight(GLuint shadow_width, GLuint shadow_height,
+PointLight::PointLight(GLfloat shadow_width, GLfloat shadow_height,
                                         GLfloat near, GLfloat far, 
                                         GLfloat red, GLfloat green, GLfloat blue,
                                         GLfloat a_intensity, GLfloat d_intensity,
