+++ BasicOpenGLEngine/BasicOpenGLEngine/PointLight.h
@@ -9,7 +9,7 @@ class PointLight :
 
 public:
     PointLight();
-    PointLight(GLuint shadow_width, GLuint shadow_height,
+    PointLight(GLfloat shadow_width, GLfloat shadow_height,
                         GLfloat near, GLfloat far,
                         GLfloat red, GLfloat green, GLfloat blue,
                         GLfloat a_intensity, GLfloat d_intensity,
